requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
azure-core==1.10.0
azure-eventhub==5.2.1
azure-eventhub-checkpointstoreblob==1.1.2
azure-eventhub-checkpointstoreblob-aio==1.1.2
azure-iot-device==2.5.0
azure-iot-hub==2.2.3
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cryptography==3.3.1
deprecation==2.1.0
idna==2.10
isodate==0.6.0
janus==0.4.0
msrest==0.6.20
multidict==5.1.0
oauthlib==3.1.0
packaging==20.8
paho-mqtt==1.5.1
pycparser==2.20
pyparsing==2.4.7
PySocks==1.7.1
requests==2.25.1
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
six==1.15.0
typing-extensions==3.7.4.3
uamqp==1.2.13
urllib3==1.25.11
yarl==1.6.3
